[refactor]#20 분리된 탑바 합체
